requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-geojson==2.12.0
django-heroku==0.3.1
django-leaflet==0.25.0
geojson==2.5.0
gunicorn==20.0.4
numpy==1.17.4
pandas==0.25.3
psycopg2==2.8.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
xlrd==1.2.0
